A function to check to values from their from data sources

This function checks and inspects the mapping and values of specified columns in a dataframe, comparing 'from' and 'to' columns if provided.

Usage

check_values(data_df, from, to = NULL, type, eval_func = as.numeric)

Arguments

data_df

A dataframe that includes the from and to variables.

from

A string or vector of strings that contain from variable names.

to

A string or vector of strings containing the name of to variables (default is NULL).

type

Either "map" or "cont" to specify the type of checking to be performed.

eval_func

A function call for the type of check that should be performed on "cont" to variables (default is as.numeric).

Value

An invisible dataframe with the results of the check.

Examples

# Assuming you have a DataFrame called 'my_data' with 'from_col' and 'to_col' columns
check_values(
  data_df = my_data,
  from = "from_col",
  to = "to_col",
  type = "map"
)
#> Error: object 'my_data' not found