Compare 2 dataframes

This function compares two dataframes by analyzing their numeric, factor, character, binary, and date fields. It returns a list of dataframes with the results of the comparison.

Usage

compare_dataframes(
  oldData = NULL,
  newData = NULL,
  suffixTerm = "",
  indOutcomes = c(""),
  surgCol = "surgeon"
)

Arguments

oldData

A dataframe representing the old data.

newData

A dataframe representing the new data. If NULL, the function will only analyze the oldData.

suffixTerm

A character string representing the suffix term.

indOutcomes

A character vector representing indent outcomes (not sure if this is accurate).

surgCol

A character string representing the surgeon column name.

Value

A list of dataframes containing the results of the comparison.

Examples

old_data <- data.frame(...)
#> Error: '...' used in an incorrect context
new_data <- data.frame(...)
#> Error: '...' used in an incorrect context
compare_dataframes(oldData = old_data, newData = new_data)
#> Error: object 'new_data' not found