Filter Cases from a Data Frame

This function filters cases from a data frame based on a given logical expression. It also groups the data by specified variables if required. NA values in the selected columns are also excluded during the filtering.

Usage

filter_cases(
  data_df,
  select,
  logic,
  data_path = file.path(getwd(), "secure_data", PROJECT_DICT$data_date),
  group_by_vars = NULL,
  rds = NULL,
  remove = TRUE
)

Arguments

data_df

A data frame to be filtered.

select

A character vector of variable names used in the logical expression.

logic

A character string representing the logical expression to filter cases.

data_path

A character string specifying the file path for saving the RDS file (default: "secure_data/<data_date>").

group_by_vars

A character vector of variable names to group by (default: NULL).

rds

A character string specifying the filename for the RDS file storing removed cases.

remove

A logical value indicating whether to remove cases that don't meet the logical expression (default: TRUE).

Value

An invisible data frame with either the cases that meet the logical expression (if remove = TRUE) or the ungrouped input data (if remove = FALSE).

Examples

# Filter cases where age is greater than or equal to 18 and not NA
filtered_data <- filter_cases(data_df = my_data,
select = "age",
logic = "age >= 18",
rds = "cases_removed")
#> Error: object 'my_data' not found