Log Information About Comments and TODOs

This function logs information about comments and TODOs added in a file, storing metadata in a specified data path.

Usage

line_log(
  type,
  item,
  current_file = get_current_file(),
  data_path = file.path(getwd(), "data", PROJECT_DICT$data_date)
)

Arguments

type

A character string representing the type of annotation, either "COMMENT" or "TODO".

item

A character string or object containing the comment or TODO line(s).

current_file

A character string representing the name of the current file being edited. Default is the current file.

data_path

A character string representing the path to the directory where log files should be saved. Default is a "secure_data" subfolder in the working directory.

Value

NULL. This function is called for its side effect of saving log files.

Examples

# Log a comment
line_log("COMMENT", "This is a comment.")
#> Error: Failed to evaluate glue component {data_path}
#> Caused by error:
#> ! object 'PROJECT_DICT' not found

# Log a TODO with a custom current_file and data_path
line_log("TODO", "This is a TODO.", "my_file.R", "custom_data_path")
#> Warning: cannot open compressed file 'custom_data_path/Metadata/Line Logs/TODO_203449.rds', probable reason 'No such file or directory'
#> Error in gzfile(file, mode): cannot open the connection