Function to run the "function"/"scripts" that populate main.R

Wrapper function for a do.call() command applied to script function. Benefit is de-clutter the calls. Drawback is debugging more convoluted.

Usage

run_file(
  file,
  args = NULL,
  ignore_warnings =
    "NAs introduced by coercion|failed to parse|no non-missing arguments to max; returning -Inf|no non-missing arguments to min; returning Inf"
)

Arguments

file

A function/script call

args

A list of parameters that should be specified prior to running the function/script specified in "file"

ignore_warnings

For now, a default logic to muffle a few routine "warnings" that are likely best to prevent from clogging up the output in the main.R script (NAs values created). Can set to "" to override this functionality.

Value

return_list