Save issue data to a specified path

This function saves issue data to a specified path and prints the saved data to the console. It takes a data frame, an RDS object, and optional data and issues paths as inputs.

Usage

save_issue(
  data_df,
  rds,
  data_path = file.path(getwd(), "secure_data", PROJECT_DICT$data_date),
  issues_path = file.path(getwd(), "secure_data", PROJECT_DICT$data_date, "metadata",
    "issues")
)

Arguments

data_df

A data frame containing issue data to be saved

rds

An RDS object to be used for saving data

data_path

An optional character string specifying the data path. Defaults to 'file.path(getwd(), "secure_data", PROJECT_DICT$data_date)'

issues_path

An optional character string specifying the issues path. Defaults to 'file.path(getwd(), "secure_data", PROJECT_DICT$data_date, 'metadata', 'issues')'

Value

An invisible copy of the saved data frame

Examples

# Save issue data to the default paths
save_issue(data_df, rds)
#> Error: Failed to evaluate glue component {data_path}
#> Caused by error:
#> ! object 'PROJECT_DICT' not found

# Save issue data to custom paths
save_issue(data_df, rds, "custom_data_path", "custom_issues_path")
#> Error: Failed to evaluate glue component {rds}
#> Caused by error:
#> ! object 'rds' not found