Parse a dataframe and compute statistics for different variable types

This function takes a dataframe as input and computes statistics for different variable types, including dates, binary, characters, factors, and numeric variables. It can also compute statistics per surgeon and per year if specified.

Usage

parse_function(
  parse_df,
  suffixTerm = "",
  indOutcomes = c(""),
  surgCol = "surgeon",
  addYears = FALSE
)

Arguments

parse_df

Dataframe to parse

suffixTerm

Suffix term (default: "")

indOutcomes

A character vector of individual outcomes to compare per surgeon (default: c(""))

surgCol

A character string specifying the surgeon column (default: "surgeon")

addYears

A boolean indicating whether to add year values (default: FALSE)

Value

A list of dataframes containing the computed statistics for the different variable types

Examples

# Assuming you have a dataframe called df
parsed_data <- parse_function(df)
#> Error in UseMethod("ungroup"): no applicable method for 'ungroup' applied to an object of class "function"
# To include year values and individual outcomes per surgeon:
parsed_data <- parse_function(df, indOutcomes = c("outcome1", "outcome2"), addYears = TRUE)
#> Error in UseMethod("ungroup"): no applicable method for 'ungroup' applied to an object of class "function"